requirements.txt
ntwork>=0.9.0
lark-oapi>=1.0.0
openai>=1.0.0
anthropic>=0.20.0
pydantic>=2.0
pydantic-settings>=2.0
pyyaml>=6.0
python-dotenv>=1.0
loguru>=0.7
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.25
aiohttp>=3.9
